requirements.txt
Flask
Flask-Login
Flask-SocketIO
SQLAlchemy
Flask-SQLAlchemy
PyMySQL
Flask-Bcrypt
Flask-Migrate
python-dotenv
pdfminer.six
pytesseract
pdf2image
fpdf2
python-docx
google-generativeai
Pillow
gunicorn
eventlet
Flask-Cors
Flask-Session
Flask-WTF
WTForms
Email-Validator
itsdangerous
Jinja2
MarkupSafe
Werkzeug
click
colorama
blinker
Flask-Bootstrap
